Deploy

+++ Server.js
@@ -21,6 +21,11 @@ require('dotenv').config({path: __dirname + '/util/.env'});
 require('dotenv').config({path: __dirname + '/script/.env'});
 require('dotenv').config({path: __dirname + '/mongodb/.env'});
 
+//Email
+EmailTemplate = require('email-templates').EmailTemplate,
+path = require('path'),
+Promise = require('bluebird');
+
 // Cookies
 function parseCookies (request) {
     var list = {},
@@ -35,7 +40,7 @@ function parseCookies (request) {
 }
 
 // Static pages
-//server.use(express.static('htmls'));
+server.use('/static', express.static('public'));
 
 // Oauth2 setup
 // server.use(oauth2.inject());
@@ -57,18 +62,14 @@ function parseCookies (request) {
 
 // Error handling
 function sendErrorMessage(code, request, response) {
-    let rawdata = fs.readFileSync(serverUtils.fetchFile(Constants.SCRIPT_ERRORS_PATH));
-    let parsedData = JSON.parse(rawdata);
+    let error;
     if (typeof(code)==='string') {
-        var name = code
-        code = parsedData.length - 1
-        while (1) {
-            if (parsedData[code]["Name"] == name) break;
-            code -= 1;
-            if (code < 0) { code = 1; break; }
-        }
+        error = serverUtils.findErrorByName(code)
+    } else {
+        let rawdata = fs.readFileSync(serverUtils.fetchFile(Constants.SCRIPT_ERRORS_PATH));
+        let parsedData = JSON.parse(rawdata);
+        error = parsedData[code];
     }
-    let error = parsedData[code];
     let errorData = error["Data"][request.header("Locale") != null ? request.header("Locale") : Constants.DEFAULT_LOCALE];
     let thisErr = {
         "Error": errorData["PrettyName"],
@@ -78,36 +79,59 @@ function sendErrorMessage(code, request, response) {
     response.status(error["HttpReturn"]).header("Content-Type", "application/json").send(JSON.stringify(thisErr));
 }
 
+//Email template
+function loadTemplate(templateName, contexts){
+    let template = new EmailTemplate(path.join(__dirname, '/templates', templateName));
+    return Promise.all([contexts].map((context) => {
+        return new Promise((resolve, reject) => {
+            template.render(context, (err, result) => {
+                if (err) reject(err);
+                else resolve({
+                    email: result,
+                    context,
+                });
+            });
+        });
+    }));
+}
+
 // =================================== Requests ===================================
 
 server.post(Constants.CREATE_ACCOUNT_REQUEST, async function(req, res) {
     let data = req.body;
     let authToken = req.token;
-
-    let findResult = await mongo.db.collection('users').findOne({'email':data['email']});
+    
+    let findResult = await mongo.db.collection('users').findOne({[Constants.USER_PRIMARY_KEY]:data[Constants.USER_PRIMARY_KEY]});
     if (findResult) {
-        logger.info("Account requested for email " + data['email'] + " already in use");
-        sendErrorMessage(3, req, res); //TODO find a better way to reply
-        return
+        logger.info("Account requested for PK ("  + Constants.USER_PRIMARY_KEY + ") " + data[Constants.USER_PRIMARY_KEY] + " already in use");
+        sendErrorMessage("PrimaryKeyInUse", req, res); //TODO find a better way to reply
+        return 
     }
-    var newUser = new userModel.User(data['email'], data['name'], data['password']).toJSON();
+    var newUser = new userModel.User(
+        data['email'], 
+        data['name'], 
+        serverUtils.saltAndHashPassword(
+            data['email'], 
+            data['password']
+        )
+    ).toJSON();
     newUser['authToken'] = serverUtils.generateToken(32);
-    logger.info("Creating user : ", newUser);
+    logger.info("Creating user : ", newUser[Constants.USER_PRIMARY_KEY]);
     let result = await mongo.db.collection('pendingUsers').insertOne(newUser);
     if (result == null) {
         sendErrorMessage(1, req, res);
     } else {
         sendErrorMessage(0, req, res); //TODO find a better way to reply
         //TODO des-gambiarrar esse processo de enviar email
-        mailer.sendMail({
-            from: Constants.SOURCE_EMAIL_ADDRESS,
-            to: newUser['email'],
-            subject: 'Street analyzer account validation',
-            text: 'That was easy!\n' +
-                'Now just click on this link to validate your account: ' +
-                serverUtils.serverUrl +
-                Constants.VERIFY_ACCOUNT_REQUEST +
-                '?token='+newUser['authToken']
+        loadTemplate('validation', newUser).then((results) => {
+            return Promise.all(results.map((result) =>{         
+                mailer.sendMail({
+                    from: Constants.SOURCE_EMAIL_ADDRESS,
+                    to: newUser['email'],
+                    subject: 'Street analyzer account validation',
+                    html: result.email.html
+                });
+            }));
         });
     }
 });
@@ -115,6 +139,10 @@ server.post(Constants.CREATE_ACCOUNT_REQUEST, async function(req, res) {
 server.get(Constants.VERIFY_ACCOUNT_REQUEST, async function(req, res) {
     let query = req.query;
     let authToken = query.token;
+    if (authToken == null) {
+        sendErrorMessage("MalformedToken", req, res);
+        return;
+    }
 
     let auth = await mongo.db.collection('pendingUsers').findOneAndDelete({'authToken':authToken});
     if (auth == null) {
@@ -131,12 +159,14 @@ server.get(Constants.VERIFY_ACCOUNT_REQUEST, async function(req, res) {
 server.post(Constants.AUTH_REQUEST, async function(req, res) {
     let data = req.body;
     // TODO use SHA256 of password
-    let authResult = await mongo.createSession(data.user, data.pass);
+    let authResult = await mongo.createSession(data[Constants.USER_PRIMARY_KEY], data[Constants.USER_PASSWORD_KEY]);
     logger.debug("Authentication result for " + JSON.stringify(data) + " is " + String(authResult))
-    if (authResult) {
-        res.status(200).header("Content-Type", "application/json").send(JSON.stringify({[Constants.AUTH_TOKEN_KEY]:authResult}));
+    if (typeof(authResult) === 'string') {
+        let userData = mongo.getUser(data[Constants.USER_PRIMARY_KEY]);
+        userData[Constants.AUTH_TOKEN_KEY] = authResult;
+        res.status(200).header("Content-Type", "application/json").send(JSON.stringify(userData));
     } else {
-        sendErrorMessage("InvalidCredentials", req, res);
+        sendErrorMessage(authResult['id'], req, res);
     }
 });
 
@@ -156,7 +186,7 @@ server.get(Constants.DEAUTH_REQUEST, async function(req, res) {
         return;
     }
 
-    sendErrorMessage(0, req, res);
+    sendErrorMessage(0, req, res); //TODO find a better way to replyc
 });
 
 server.get(Constants.QUALITY_OVERLAY_REQUEST, function(req, res) {
